[FSDP()][4/N] Refactor `ShardingStrategy` enum
[ghstack-poisoned]
